+++ CHANGELOG.md
@@ -4,6 +4,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). See [MAINTAINERS.md](./MAINTAINERS.md)
 for instructions to keep up to date.
 
+## Unreleased
+
+### Added
+
+* added `tools compare-blocks-rpc` command to perform a validation between firehose blocks and blocks+trx+logs fetched from an RPC endpoint
+
 ## v1.4.12
 
 ### Highlights